Basic Organization of the folder
This is an initialization for the End of study project.
In the report folder we will be adding more and more elements as we go.
And in the conception folder we will be storing every Diagrams that we created through this project, which contain a documentation folder that will be storing documentation about every aspects of the project form the textual description for the Actors and Usecases etc..
The controller folder will hold our backend logic,
The Web Application folder will hold our Frontend Application/s deppending if we are going to need a separate Frontend for other purpusose.